requirements.txt
# Core dependencies
streamlit==1.32.0
pandas==2.1.4
plotly==5.17.0
numpy==1.24.3
python-dotenv==1.0.0

# Data processing
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Development and testing
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Optional: Spotify API (limited functionality)
spotipy==2.23.0

# Optional: Jupyter for notebooks
jupyter==1.0.0
ipykernel==6.25.2 